# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bob Engine \u2014 FIR Intelligence & Crime Pattern Detector",
    "description": "Surfaces pattern correlations across FIR records using deterministic entity resolution and MO-similarity clustering.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/firs": {
      "get": {
        "tags": [
          "FIRs"
        ],
        "summary": "List Firs",
        "operationId": "list_firs_firs_get",
        "parameters": [
          {
            "name": "crime_category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crime Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/firs/{fir_id}": {
      "get": {
        "tags": [
          "FIRs"
        ],
        "summary": "Get Fir",
        "operationId": "get_fir_firs__fir_id__get",
        "parameters": [
          {
            "name": "fir_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Fir Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/summary": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "Dashboard Summary",
        "operationId": "dashboard_summary_dashboard_summary_get",
        "parameters": [
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offenders": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "List Offenders",
        "operationId": "list_offenders_offenders_get",
        "parameters": [
          {
            "name": "syndicate_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Syndicate Only"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/suspects/{suspect_id}": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "Get Dossier",
        "operationId": "get_dossier_suspects__suspect_id__get",
        "parameters": [
          {
            "name": "suspect_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Suspect Id"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/graph": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "Get Graph",
        "operationId": "get_graph_graph_get",
        "parameters": [
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stations/all": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "List All Stations",
        "description": "Unscoped \u2014 feeds the Topbar station picker for all roles.",
        "operationId": "list_all_stations_stations_all_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/stations": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "List Stations",
        "operationId": "list_stations_stations_get",
        "parameters": [
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/heatmap": {
      "get": {
        "tags": [
          "Intelligence"
        ],
        "summary": "Get Heatmap",
        "operationId": "get_heatmap_heatmap_get",
        "parameters": [
          {
            "name": "crime_category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crime Category"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/detect": {
      "post": {
        "tags": [
          "Intelligence"
        ],
        "summary": "Run Detect",
        "description": "Manually trigger syndicate detection. Mainly for debugging.",
        "operationId": "run_detect_detect_post",
        "parameters": [
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "summary": "List Alerts",
        "operationId": "list_alerts_alerts_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-User-Role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-User-Role"
            }
          },
          {
            "name": "X-User-Station",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Station"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}